Add files via upload
Write a Python program that…
• Reads the file stephen_king_adaptations.txt
• Copies all the content from that file to a list called stephen_king_adaptations_list
• Establishes a connection with a new SQLite database called 
stephen_king_adaptations.db
• Creates a table called stephen_kind_adaptations_table with the column names 
movieID, movieName, movieYear, imdbRating.
• Takes the content from stephen_king_adaptations_list and inserts it into the table 
stephen_king_adaptations_table
• Gives the user the option to search for movies in the database, based on the following 
parameters. This should be presented to the user on a loop, until the user picks the 
option to stop:
1. Movie name
2. Movie year
3. Movie rating
4. STOP
• If the user selects Option 1, then ask the user for the name of the movie to be searched 
in the database. 
1. If the movie is found, the program should display all the details of that movie, 
which includes the name, the year, and the rating.
2. If the movie is not found, then the message “No such movie exists in our 
database” should be displayed
• If the user selecs Option 2, then ask the user for a year and the program should return 
all the movie details (like above) from the database that were released in that year
1. If no movie is found for that year, then the message “No movies were found 
for that year in our database.”
• If the user selects Option 3, then ask the user for the rating for and above which the 
movies can be returned. For example, if the user enters 5, then only those movies that 
have the IMDB rating either 5 or more than 5 will be returned.
1. If no movies fall within the provided rating limit, then the message “No movies 
at or above that rating were found in the database.
• If the user selects OPtion 4, then the program should terminate.